move test_critical_density
