+++ astropy/cosmology/flrw/tests/test_lambdacdm.py
@@ -561,3 +561,28 @@ def test_neg_distmod():
         tcos.luminosity_distance([50, 100]), [16612.44047622, -46890.79092244] * u.Mpc
     )
     assert u.allclose(tcos.distmod([50, 100]), [46.102167189, 48.355437790944] * u.mag)
+
+
+@pytest.mark.skipif(not HAS_SCIPY, reason="test requires scipy")
+def test_critical_density():
+    from astropy.constants import codata2014
+
+    # WMAP7 but with Omega_relativistic = 0
+    # These tests will fail if astropy.const starts returning non-mks
+    #  units by default; see the comment at the top of core.py.
+    # critical_density0 is inversely proportional to G.
+    tcos = FlatLambdaCDM(70.4, 0.272, Tcmb0=0.0)
+
+    fac = (const.G / codata2014.G).to(u.dimensionless_unscaled).value
+    assert u.allclose(
+        tcos.critical_density0 * fac, 9.309668456020899e-30 * (u.g / u.cm**3)
+    )
+    assert u.allclose(tcos.critical_density0, tcos.critical_density(0))
+    assert u.allclose(
+        tcos.critical_density([1, 5]) * fac,
+        [2.70352772e-29, 5.53739080e-28] * (u.g / u.cm**3),
+    )
+    assert u.allclose(
+        tcos.critical_density([1.0, 5.0]) * fac,
+        [2.70352772e-29, 5.53739080e-28] * (u.g / u.cm**3),
+    )
